Final commit. Support for multiple files. TODO: fix bug with memmove Figure out why there is a memory leak for compilation with -O3 and above

+++ symbol.h
@@ -16,7 +16,7 @@ typedef struct {
 typedef struct {
 	char *sym;
 	Status status;
-	int module_id;
+	short module_id;
 } Uselist_Symbol;
 
 Symbol *read_symbol(FILE *stream);
